Added option parsing for more useful flexible command line use
This included significant modifications to some of the script structure.
- Added __init__.py to parser to make a python module for cleaner
importing - Added a main() method to handle options parsing and control
flow. - Moved output function to a method of the "Title" class for
easier additions of output formats and to have better support for future
implementation of limiting output to a single title based on optional
queries.